feat: Introduce CLI, centralize NFL season/week configuration, and update workflow.
